Commit de todas as alterações
