--- Backlog/.editorconfig
@@ -1,4 +1,0 @@
-﻿[*.cs]
-
-# CS8618: Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-dotnet_diagnostic.CS8618.severity = silent
