--- Backlog/Backlog/Data/DataContext.cs
@@ -1,10 +1,0 @@
-﻿namespace Backlog.Data
-{
-    public class DataContext : DbContext
-    {
-        public DataContext(DbContextOptions<DataContext> options) : base(options) { }
-        public DbSet<User> User { get; set; }
-        public DbSet<Ticket> Ticket { get; set; }
-        public DbSet<Historic> Historic { get; set; }
-    }
-}
