--- Backlog/Backlog/Services/Implamentations/TicketServiceImplementation.cs
@@ -1,35 +1,0 @@
-﻿namespace Backlog.Services.Implamentations
-{
-    public class TicketServiceImplementation : ITicketService
-    {
-        public Ticket Create(Ticket user)
-        {
-            throw new NotImplementedException();
-        }
-
-        public void Delete(long id)
-        {
-            throw new NotImplementedException();
-        }
-
-        public List<Ticket> FindAll()
-        {
-            throw new NotImplementedException();
-        }
-
-        public Ticket FindyByEmail(string email)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Ticket FindyByID(long id)
-        {
-            throw new NotImplementedException();
-        }
-
-        public Ticket Update(User user)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
