+++ PlaceHolder/PlaceHolder.sln
@@ -1,0 +1,30 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.32210.238
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PlaceHolder", "PlaceHolder\PlaceHolder.csproj", "{D23237B1-3359-4ECC-A247-4A498F8B95C2}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{528460E4-5798-453D-BADC-45F0DFCF4831}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D23237B1-3359-4ECC-A247-4A498F8B95C2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D23237B1-3359-4ECC-A247-4A498F8B95C2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D23237B1-3359-4ECC-A247-4A498F8B95C2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D23237B1-3359-4ECC-A247-4A498F8B95C2}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {804845C6-7A8B-46E0-9D11-BD6E9D20A5B2}
+	EndGlobalSection
+EndGlobal
