+++ PlaceHolder/PlaceHolder/Repositories/IHistoricRepository.cs
@@ -1,0 +1,17 @@
+﻿namespace PlaceHolder.Repositories.Implamentations
+{
+    public interface IHistoricRepository
+    {
+        Historic? Create(Historic historic);
+
+        Historic? Update(Historic historic);
+
+        void Delete(long id);
+
+        Historic? FindByID(long id);
+
+        List<Historic> FindAll();
+
+        Boolean IsExist(Historic historic);
+    }
+}
