+++ PlaceHolder/PlaceHolder/Services/IUserService.cs
@@ -1,0 +1,19 @@
+﻿using PlaceHolder.DTOs;
+
+namespace PlaceHolder.Services.Implamentations
+{
+    public interface IUserService
+    {
+        User Create(UserDTO user);
+
+        User? Update(User user);
+
+        void Delete(string email);
+
+        User? FindByID(long id);
+
+        User? FindByEmail(string email);
+
+        List<User> FindAll();
+    }
+}
